Added more pre-processing to make runtime quicker. Also, seperate functions into relevant files.
