+++ SingletonDeckState.py
@@ -1,0 +1,14 @@
+from deck_setup import connect_to_elgato
+class SingletonDeckState:
+    _instance = None
+
+    def __new__(cls):
+        if cls._instance is None:
+            cls._instance = super(SingletonDeckState, cls).__new__(cls)
+            # Initialize your variables here
+            cls._instance.current_page = 0
+            cls._instance.current_row = 0
+            cls._instance.process_input = True
+            cls._instance.deck = connect_to_elgato()
+
+        return cls._instance
